Moves logic for Tab and TabItem into separate files
